refactor: shared CommonInputProps and shared FormDebounceOptions
