+++ packages/bootstrap-vue-next/src/types/index.ts
@@ -11,7 +11,7 @@ export type * from './ColorTypes'
 export type * from './CommonTypes'
 export type * from './ComponentOrchestratorTypes'
 export type * from './ComponentProps'
-export type {FormCommonInputProps} from './FormCommonInputProps'
+export type {CommonInputProps, FormDebounceOptions} from './FormCommonInputProps'
 export type {InputType} from './InputType'
 export type {LinkTarget} from './LinkTarget'
 export type {LiteralUnion} from './LiteralUnion'
